wah4h-backend/requirement.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
requests